Fix three signed channels the manual check missed
idleAngleCorr, cam1Angle and cam2Angle are sbyte/sword in Ecumaster's format
definition and were read unsigned, so every negative reading came out as a
large positive one.

idleAngleCorr is the one that matters: idle ignition correction goes negative
whenever the ECU pulls timing to hold the idle target, which at idle is most
of the time. A -3 degree correction read as +126. The upcoming IDLE page shows
this channel, which is how it was found.

The cam pair shows why the first pass missed them. cam1AngleTarget and
cam2AngleTarget sit on the next two lines and were both caught; reading the
group as done left the two actual angles unfixed.

So the check is now a script rather than an eye: tools/edl_check_storage.py
re-derives storage and divider for all 195 channels from the XML and compares
them against what parseFrame() does. All 195 now agree, with fcProbability
still exempt for the reason the README gives.

+++ lib/EDLSerial/src/EDLSerial.cpp
@@ -64,7 +64,7 @@ bool EDLSerial::parseFrame(uint8_t *data, EDLFrame &frame) {
   frame.idleTarget=data[33] | (data[34] << 8);
   frame.idleDC=(uint8_t)data[35];
   frame.idleStep=(uint8_t)data[36];
-  frame.idleAngleCorr=data[37] / 2.0f;
+  frame.idleAngleCorr=(int8_t)data[37] / 2.0f;
   frame.idleControlActive=(uint8_t)data[38];
   frame.curentPIDCorrection=(int8_t)data[39];
   frame.dwellTime=data[40] / 20.0f;
@@ -137,8 +137,8 @@ bool EDLSerial::parseFrame(uint8_t *data, EDLFrame &frame) {
   frame.iatFuelCorr=(uint8_t)data[121];
   frame.parametricOutput3=(uint8_t)data[122];
   frame.parametricOutput4=(uint8_t)data[123];
-  frame.cam1Angle=(data[124] | (data[125] << 8)) / 2.0f;
-  frame.cam2Angle=(data[126] | (data[127] << 8)) / 2.0f;
+  frame.cam1Angle=(int16_t)(data[124] | (data[125] << 8)) / 2.0f;
+  frame.cam2Angle=(int16_t)(data[126] | (data[127] << 8)) / 2.0f;
   frame.cam1ValveDC=(uint8_t)data[128];
   frame.cam2ValveDC=(uint8_t)data[129];
   frame.cam1AngleTarget=(int16_t)(data[130] | (data[131] << 8)) / 2.0f;
